Organize the main functions
