+++ layer/__init__.py
@@ -1,7 +1,6 @@
 from .context import Context  # noqa
 from .contracts.datasets import Dataset  # noqa
-from .contracts.logged_data import Markdown  # noqa
-from .contracts.logged_data import Image, Video  # noqa
+from .contracts.logged_data import Image, Markdown, Video  # noqa
 from .contracts.models import Model  # noqa
 from .contracts.projects import Project  # noqa
 from .contracts.runs import Run  # noqa
@@ -23,19 +22,19 @@ from .decorators.assertions import (  # noqa
 from .flavors.custom import CustomModel  # noqa
 from .global_context import current_project_full_name  # noqa
 from .logged_data.callbacks import KerasCallback, XGBoostCallback  # noqa
-from .main import clear_cache  # noqa
-from .main import get_dataset  # noqa
-from .main import get_model  # noqa
-from .main import get_version  # noqa
-from .main import init  # noqa
-from .main import log  # noqa
-from .main import login  # noqa
-from .main import login_as_guest  # noqa
-from .main import login_with_access_token  # noqa
-from .main import login_with_api_key  # noqa
-from .main import logout  # noqa
-from .main import run  # noqa
-from .main import show_api_key  # noqa
+from .main.asset import get_dataset, get_model  # noqa
+from .main.auth import (  # noqa
+    login,
+    login_as_guest,
+    login_with_access_token,
+    login_with_api_key,
+    logout,
+    show_api_key,
+)
+from .main.cache import clear_cache  # noqa
+from .main.log import log  # noqa
+from .main.run import init, run  # noqa
+from .main.version import get_version  # noqa
 from .pandas_extensions import Arrays, Images, _register_type_extensions  # noqa
